requirements.txt
anyascii==0.2.0
asgiref==3.3.4
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-filter==2.4.0
django-modelcluster==5.1
django-snippet-image==0.1.3
django-taggit==1.4.0
django-treebeard==4.5.1
djangorestframework==3.12.4
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
html5lib==1.1
idna==2.10
l18n==2020.6.1
openpyxl==3.0.7
Pillow==8.2.0
pytz==2021.1
requests==2.25.1
six==1.16.0
snippet-image==0.1.7
soupsieve==2.2.1
sqlparse==0.4.1
tablib==3.0.0
telepath==0.2
typing-extensions==3.10.0.0
urllib3==1.26.5
wagtail==2.13.2
webencodings==0.5.1
Willow==1.4
xlrd==2.0.1
XlsxWriter==1.4.3
xlwt==1.3.0
